% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.est.R
\name{cum.haz}
\alias{cum.haz}
\title{Calulate the Values of the cumulative Hazard at Fixed Times}
\usage{
cum.haz(est, t.eval = c(0.25, 0.75))
}
\arguments{
\item{est}{an object of the class \code{plac.fit}.}

\item{t.eval}{time points at which the Lambda(t) is evaluated (for both conditional apporach and the PLAC estimator).}
}
\value{
a list containing the estiamtes and SEs of Lambda(t) for both conditional apporach and the PLAC estimator.
}
\description{
Calulate the Values of the cumulative Hazard at Fixed Times
}
\examples{
dat1 = sim.ltrc(n = 100)$dat
est = PLAC(ltrc.formula = Surv(As, Ys, Ds) ~ Z1 + Z2,
     ltrc.data = dat1, td.type = "none")
H = step.L(est, t.eval = seq(0.1, 0.9, 0.1))
H$L
H$se.L
}

