% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen.R
\name{get.Cmax}
\alias{get.Cmax}
\title{Obtain the required Cmax for a specific censoring proportion (the inverse of
\code{sim.pc()}).}
\usage{
get.Cmax(pc = 0.5, l = 0.1, u = 10, n = 500, I = 500,
  distr.T = "weibull", distr.A = "weibull", shape.T = 2, scale.T = 1,
  shape.A = 1, scale.A = 5, p.A = 0.3, M = 100)
}
\arguments{
\item{pc}{numeric. Proportion of censoring.}

\item{l}{numeric. The lower bound of the searching interval.}

\item{u}{numeric. The upper bound of the searching interval.}

\item{n}{An integer for sample size.}

\item{I}{An integer iteration number.}

\item{distr.T}{A character for the baseline survival time (T*) distribution
("exp" or "weibull").}

\item{distr.A}{A character for the baseline truncation time (A*) distribution
("weibull", "unif" (Length-Biased Sampling) or "binomial").}

\item{shape.T}{The shape parameter for the Weibull distribution of T*.}

\item{scale.T}{The scale parameter for the Exponential/Weibull distribution
of T*.}

\item{shape.A}{The shape parameter for the Weibull distribution of A*.}

\item{scale.A}{The scale parameter for the Weibull distribution of A*.}

\item{p.A}{The success probability for the binomial distribution of A*.}

\item{M}{integer. The maximal iteration for \code{uniroot}.}
}
\description{
Obtain the required Cmax for a specific censoring proportion (the inverse of
\code{sim.pc()}).
}

