% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.est.R
\name{PLAC}
\alias{PLAC}
\title{Calculate the PLAC estimator when a time-dependent indicator presents}
\usage{
PLAC(formula, data, id.var, td.var, type = "none", tjump, init.val = NULL,
  max.iter = 100, ...)
}
\arguments{
\item{formula}{a formula of of the form \code{Surv(A, X, D) ~ Z}, where \code{Z} only include the time-invariate covariates.}

\item{data}{a data.frame of the LTRC dataset including the responses, time-invariate covariates and the jump times for the time-depnencent covariate.}

\item{id.var}{a name of the subject id in \code{data}.}

\item{td.var}{a name of the time-dependent covariate in the output.}

\item{type}{the type of the time-dependent covariate. Either one of \code{c("none", "independent", "post-trunc", "pre-post-trunc")}. See Details.}

\item{tjump}{a name of the jump time variable in \code{data}.}

\item{init.val}{a list of the initial values of the coefficients and the baseline hazard function for the PLAC estimator.}
}
\value{
a list of model fitting results for both conditional approach and the PLAC estimators.
}
\description{
Both a conditional approach Cox model and a pairwise likelihood augmented estimator are fitted and the corresponding results are returned in a list.
}
\details{
The formula should have the same format as used in \code{coxph()}, where \code{A} is the truncation time (\code{tstart}), \code{X} is the survival time (\code{tstop}) and \code{D} is the status indicator (\code{event}).
\code{type} is used to determine which \code{C++} function (either \code{PLAC_TV} or \code{PLAC_TVR}) to call. \code{"independent"} will invoke the former, whereas \code{"post-trunc"} and \code{"pre-post-trunc"} will invoke the latter.
For \code{type = "post-trunc"}, the pre-truncation values for the time-dependent covariate will be set to be zero for all subjects.
}
\references{
Wu, F. Kim, S. and Li, Y. "A Pairwise Likelihood Augmented Estimator for Left-Truncated Data with Time-Dependent Covariates." (\emph{in preparation})

Wu, F., Kim, S., Qin, J., Saran, R. and Li, Y. (2015) "A Pairwise-Likelihood Augmented Estimator for the Cox Model Under Left-Truncation." (Submitted to \emph{Journal of American Statistical Association}.)
}

