% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.est.R
\name{plactv}
\alias{plactv}
\title{Calculate the PLAC estimator when a time-dependent indicator presents}
\usage{
plactv(formula, data, id.var, td.var, tjump, init.val = NULL,
  max.iter = 100, ...)
}
\arguments{
\item{formula}{a formula of of the form \code{Surv(A, X, D) ~ Z} including the time-invariate covariates only.}

\item{id.var}{a name of the subject id in \code{data}.}

\item{td.var}{aname of the time-dependent covariate in the output.}

\item{tjump}{a name of the jump time variable in \code{data}.}

\item{init.val}{a list of the initial values of the coefficients and the baseline hazard function for the PLAC estimator.}

\item{a}{data.frame of the LTRC dataset including the responses, time-invariate covariates and the jump times for the time-depnencent covariate.}
}
\value{
a list of model fitting results for both conditional approach and the PLAC estimators.
}
\description{
Both a conditional approach Cox model and a pairwise likelihood augmented estimator are fitted and the corresponding results are returned in a list.
}

