% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PwInd}
\alias{PwInd}
\title{Generate truncation-pair indicators}
\usage{
PwInd(X, W)
}
\arguments{
\item{X}{the response matrix (As, Xs, Ds).}

\item{W}{the ordered observed event times.}
}
\value{
the truncation-pair indicators of the form I(w_k <= A_i)
\itemize{
\item I(w_k <= XA_j).
}
}
\description{
Generate truncation-pair indicators
}
