% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen.R
\name{sim.data}
\alias{sim.data}
\title{Generate left-truncated (and right-cencored) data from the Cox model.}
\usage{
sim.data(n = 200, b = c(1, 1), distr.T = "weibull", shape.T = 2,
  scale.T = 1, distr.A = "weibull", shape.A = 1, scale.A = 5,
  p.A = 0.3, Cmax = Inf, fix.seed = NA)
}
\arguments{
\item{n}{An integer for sample size.}

\item{b}{A numeric vector for true regression coefficients.}

\item{distr.T}{A character for the baseline survival time (T*) distribution
("exp" or "weibull").}

\item{shape.T}{The shape parameter for the Weibull distribution of T*.}

\item{scale.T}{The scale parameter for the Exponential/Weibull distribution
of T*.}

\item{distr.A}{A character for the baseline truncation time (A*) distribution
("weibull", "unif" (Length-Biased Sampling) or "binomial").}

\item{shape.A}{The shape parameter for the Weibull distribution of A*.}

\item{scale.A}{The scale parameter for the Weibull distribution of A*.}

\item{p.A}{The success probability for the binomial distribution of A*.}

\item{Cmax}{The upper bound of the uniform distribution of the censoring time
(C).}

\item{fix.seed}{An optional random seed for simulation.}
}
\value{
A data.frame containing the observed survival times (Ys), the
  observed truncation times (As), the event indicator (Ds) and the covariates
  (Zs).
}
\description{
Various baseline survival functions and truncation distribution are
available. Censoring rate can be designated through tuning the parameter
\code{Cmax}; \code{Cmas = Inf} means no censoring.
}
\examples{
sim.data()
sim.data(50, distr.A = "binomial", p.A = 0.8, Cmax = 5)
}

