% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen.R
\name{sim.pc}
\alias{sim.pc}
\title{Calculate the censoring rate for a survival dataset correspondes to a
specific \code{Cmax}.
Increasing the sample size and/or the iteration numbers may help make the
calculation more precise.}
\usage{
sim.pc(Cmax = 10, n = 500, I = 500, distr.T = "weibull",
  distr.A = "weibull", shape.T = 2, scale.T = 1, shape.A = 1,
  scale.A = 5, p.A = 0.3)
}
\arguments{
\item{Cmax}{The upper bound of the uniform distribution of the censoring time
(C).}

\item{n}{An integer for sample size.}

\item{I}{An integer iteration number.}

\item{distr.T}{A character for the baseline survival time (T*) distribution
("exp" or "weibull").}

\item{distr.A}{A character for the baseline truncation time (A*) distribution
("weibull", "unif" (Length-Biased Sampling) or "binomial").}

\item{shape.T}{The shape parameter for the Weibull distribution of T*.}

\item{scale.T}{The scale parameter for the Exponential/Weibull distribution
of T*.}

\item{shape.A}{The shape parameter for the Weibull distribution of A*.}

\item{scale.A}{The scale parameter for the Weibull distribution of A*.}

\item{p.A}{The success probability for the binomial distribution of A*.}
}
\value{
A decimal number of the censoring rate.
}
\description{
Calculate the censoring rate for a survival dataset correspondes to a
specific \code{Cmax}.
Increasing the sample size and/or the iteration numbers may help make the
calculation more precise.
}
\examples{
sim.pc()
sim.pc(1, n = 800, distr.A = "binomial", p.A = 0.8)
}

