% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.est.R
\name{step.L}
\alias{step.L}
\title{Calulate the Values of the cumulative Hazard at Fixed Times}
\usage{
step.L(est, t.eval = c(0.25, 0.75))
}
\arguments{
\item{est}{an object of the class \code{plac.fit}.}

\item{t.eval}{time points at which the Lambda(t) is evaluated (for both conditional apporach and the PLAC estimator).}
}
\value{
a list containing the estiamtes and SEs of Lambda(t) for both conditional apporach and the PLAC estimator.
}
\description{
Calulate the Values of the cumulative Hazard at Fixed Times
}

