% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TvInd}
\alias{TvInd}
\title{Generate time-depependent covariate indicators}
\usage{
TvInd(zeta, W)
}
\arguments{
\item{zeta}{the change point (jump time) of Z_v(t).}

\item{W}{the ordered observed event times.}
}
\value{
the time-depependent covariate of the form Z_v(t) = I(w_k > zeta).
}
\description{
Generate time-depependent covariate indicators
}
