% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TvInd1}
\alias{TvInd1}
\title{Generate time-depependent covariate indicators}
\usage{
TvInd1(Zv, W)
}
\arguments{
\item{Zv}{the jump time.}

\item{W}{the ordered observed event times.}
}
\value{
the time-depependent covariate indicators of the form I(w_k > zeta).
}
\description{
Generate time-depependent covariate indicators
}

