% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TvInd2}
\alias{TvInd2}
\title{Generate time-depependent covariate indicators}
\usage{
TvInd2(eta, zeta, W)
}
\arguments{
\item{eta}{a random variable of the Z_v(t) value before the change point.}

\item{zeta}{the change point (jump time).}

\item{W}{the ordered observed event times.}
}
\value{
the time-depependent covariate indicators of the form Z_v(t) =
eta * I(w_k <= zeta).
}
\description{
Generate time-depependent covariate indicators
}
