% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.gen.R
\name{sim.data}
\alias{sim.data}
\title{Generate left-truncated (and right-cencored) data from the Cox model.}
\usage{
sim.data(n = 200, b = c(1, 1), time.dep = FALSE, Zv.depA = FALSE,
  distr.T = "weibull", shape.T = 2, scale.T = 1, distr.A = "weibull",
  shape.A = 1, scale.A = 5, p.A = 0.3, Cmax = Inf, fix.seed = NA)
}
\arguments{
\item{n}{the sample size.}

\item{b}{a numeric vector for true regression coefficients.}

\item{time.dep}{logical, whether there is the time-dependent covariate (only
one indicator function Zv = I(t >= \eta) is supported); the default is
FALSE.}

\item{Zv.depA}{logical, whether the time-dependent covariate \code{Zv}
depends on A^* (the only form supported is Zv = I(t >= \eta + A^*)); the
default is FALSE.}

\item{distr.T}{the baseline survival time (T*) distribution ("exp" or
"weibull").}

\item{shape.T}{the shape parameter for the Weibull distribution of T*.}

\item{scale.T}{the scale parameter for the Exponential/Weibull distributiof
T*.}

\item{distr.A}{the baseline truncation time (A*) distribution: either of
\code{"weibull"} (the default), \code{"unif"} (Length-Biased Sampling),
\code{"binomial"} or \code{"dunif"}). Note: If distribution name other than
these are provided, \code{"unif"} will be used.}

\item{shape.A}{the shape parameter for the Weibull distribution of A*.}

\item{scale.A}{the scale parameter for the Weibull distribution of A*.}

\item{p.A}{the success probability for the binomial distribution of A*.}

\item{Cmax}{the upper bound of the uniform distribution of the censoring time
(C).}

\item{fix.seed}{an optional random seed for simulation.}
}
\value{
a list with a data.frame containing the observed survival times
  (\code{Xs}), the observed truncation times (\code{As}), the event indicator
  (\code{Ds}) and the covariates (\code{Zs}); a vector of certain quantiles
  of Xs (\code{taus}); the censoring proportion (\code{PC}) and the
  truncation proportiona (\code{PT}).
}
\description{
Various baseline survival functions and truncation distribution are
available. Censoring rate can be designated through tuning the parameter
\code{Cmax}; \code{Cmas = Inf} means no censoring.
}
\examples{
sim.data()
sim.data(n = 50, time.dep = TRUE, distr.A = "binomial", p.A = 0.8, Cmax = 5)
}

